Restructure app, add routing
