+++ package.json
@@ -60,6 +60,7 @@
     "react-dom": "^16.8.6",
     "react-frame-component": "^4.1.1",
     "react-json-view": "^1.19.1",
+    "react-router-dom": "^5.1.2",
     "resolve": "1.10.0",
     "sass-loader": "7.1.0",
     "semver": "6.0.0",
