+++ src/features/Metafields/MetafieldsEditor/index.js
@@ -1,0 +1,1 @@
+export { default, MetafieldsContext } from './MetafieldsEditor'
