+++ src/features/Metafields/ResourceList/ResourceList.scss
