+++ src/features/csv/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+function CSVDownloader() {
+  return <div>CSVDownloader</div>
+}
+
+CSVDownloader.propTypes = {}
+
+CSVDownloader.defaultProps = {}
+
+export default CSVDownloader
